实现内存数据库中间层，优化设置操作性能

+++ backend/app/repositories/base_repository.py
@@ -1,6 +1,7 @@
 """基础Repository类"""
 from sqlalchemy.orm import Session
 from app.core.database import SessionLocal
+from app.core.memory_database import memory_db
 
 class BaseRepository:
     """基础Repository类，提供通用的数据访问方法"""
@@ -24,6 +25,13 @@ class BaseRepository:
     
     def add(self, model):
         """添加模型实例到数据库"""
+        # 对于设置相关模型，使用内存数据库
+        model_type = self._get_model_type(model)
+        if model_type:
+            memory_db.set(model_type, model)
+            return model
+        
+        # 对于其他模型，继续使用直接数据库操作
         self.db.add(model)
         self.db.commit()
         self.db.refresh(model)
@@ -31,19 +39,53 @@ class BaseRepository:
     
     def update(self, model):
         """更新模型实例"""
+        # 对于设置相关模型，使用内存数据库
+        model_type = self._get_model_type(model)
+        if model_type:
+            memory_db.set(model_type, model)
+            return model
+        
+        # 对于其他模型，继续使用直接数据库操作
         self.db.commit()
         self.db.refresh(model)
         return model
     
     def delete(self, model):
         """删除模型实例"""
+        # 对于设置相关模型，不支持删除操作
+        model_type = self._get_model_type(model)
+        if model_type:
+            # 可以选择将其设置为None或使用默认值
+            memory_db.set(model_type, None)
+            return
+        
+        # 对于其他模型，继续使用直接数据库操作
         self.db.delete(model)
         self.db.commit()
     
     def commit(self):
         """提交事务"""
+        # 对于设置相关模型，内存数据库会自动提交
+        # 对于其他模型，继续使用直接数据库操作
         self.db.commit()
     
     def rollback(self):
         """回滚事务"""
+        # 对于设置相关模型，需要从数据库重新加载
+        # 对于其他模型，继续使用直接数据库操作
         self.db.rollback()
+    
+    def _get_model_type(self, model) -> str:
+        """获取模型类型对应的内存数据库键名"""
+        model_class_name = model.__class__.__name__
+        
+        if model_class_name == 'SystemSetting':
+            return 'system_settings'
+        elif model_class_name == 'AppSetting':
+            return 'app_settings'
+        elif model_class_name == 'VectorSetting':
+            return 'vector_settings'
+        elif model_class_name == 'NotificationSetting':
+            return 'notification_settings'
+        
+        return None
